SnapdragonCamera: Camera2 force close after focused
Handler message leaks the camera id, the camera id is set to 0 when
handleing Message.As we known, rear camera id is 0 and front camera
id is 1. Focus action runs normall when switching to rear camera, but
focus action will lead to force close when switching to fornt camera.

Change-Id: I7795b7cd0c34c256d1ad0066d13a51ed54cec67d
CRs-Fixed: 1094009